Conflicts:
	apps/src/code-studio/pd/workshop_dashboard/components/workshop_management.jsx
	apps/src/code-studio/pd/workshop_dashboard/workshop_index.jsx
	apps/test/unit/code-studio/pd/workshop_dashboard/workshop_indexTest.js
	dashboard/test/controllers/api/v1/pd/workshop_survey_report_controller_test.rb
	dashboard/test/lib/api/v1/pd/workshop_score_summarizer_test.rb
	dashboard/test/ui/features/plc/pd/facilitator_survey_results_view.feature

+++ dashboard/app/controllers/scripts_controller.rb
@@ -1,7 +1,8 @@
 class ScriptsController < ApplicationController
   include VersionRedirectOverrider
 
-  before_action :require_levelbuilder_mode, except: :show
+  before_action :require_levelbuilder_mode, except: [:show, :edit, :update]
+  before_action :require_levelbuilder_mode_or_test_env, only: [:edit, :update]
   before_action :authenticate_user!, except: :show
   check_authorization
   before_action :set_script, only: [:show, :edit, :update, :destroy]
@@ -43,6 +44,22 @@ class ScriptsController < ApplicationController
 
     @show_redirect_warning = params[:redirect_warning] == 'true'
     @section = current_user&.sections&.find_by(id: params[:section_id])&.summarize
+    sections = current_user.try {|u| u.sections.where(hidden: false).select(:id, :name, :script_id, :course_id)}
+    @sections_with_assigned_info = sections&.map {|section| section.attributes.merge!({"isAssigned" => section[:script_id] == @script.id})}
+
+    # Warn levelbuilder if a lesson will not be visible to users because 'visible_after' is set to a future day
+    if current_user && current_user.levelbuilder?
+      notice_text = ""
+      @script.lessons.each do |stage|
+        next unless stage.visible_after && Time.parse(stage.visible_after) > Time.now
+
+        formatted_time = Time.parse(stage.visible_after).strftime("%I:%M %p %A %B %d %Y %Z")
+        num_days_away = ((Time.parse(stage.visible_after) - Time.now) / 1.day).ceil.to_s
+        lesson_visible_after_message = "The lesson #{stage.name} will be visible after #{formatted_time} (#{num_days_away} Days)"
+        notice_text = notice_text.empty? ? lesson_visible_after_message : "#{notice_text} <br/> #{lesson_visible_after_message}"
+      end
+      flash[:notice] = notice_text.html_safe
+    end
   end
 
   def index
@@ -58,9 +75,9 @@ class ScriptsController < ApplicationController
   def create
     @script = Script.new(script_params)
     if @script.save && @script.update_text(script_params, params[:script_text], i18n_params, general_params)
-      redirect_to @script, notice: I18n.t('crud.created', model: Script.model_name.human)
+      redirect_to edit_script_url(@script), notice: I18n.t('crud.created', model: Script.model_name.human)
     else
-      render 'new'
+      render json: @script.errors
     end
   end
 
@@ -75,13 +92,32 @@ class ScriptsController < ApplicationController
     end
 
     @script.destroy
-    filename = "config/scripts/#{@script.name}.script"
-    File.delete(filename) if File.exist?(filename)
+    if Rails.application.config.levelbuilder_mode
+      filename = "config/scripts/#{@script.name}.script"
+      File.delete(filename) if File.exist?(filename)
+    end
     redirect_to scripts_path, notice: I18n.t('crud.destroyed', model: Script.model_name.human)
   end
 
   def edit
+    beta = params[:beta].present?
+    if @script.script_levels.any?(&:has_experiment?)
+      beta = false
+      beta_warning = "The beta Script Editor is not available, because it does not support level variants with experiments."
+    end
     @show_all_instructions = params[:show_all_instructions]
+    @script_data = {
+      script: @script ? @script.summarize_for_edit : {},
+      i18n: @script ? @script.summarize_i18n : {},
+      beta: beta,
+      betaWarning: beta_warning,
+      levelKeyList: beta && Level.key_list,
+      lessonLevelData: @script_file,
+      locales: options_for_locale_select,
+      script_families: ScriptConstants::FAMILY_NAMES,
+      version_year_options: Script.get_version_year_options,
+      is_levelbuilder: current_user.levelbuilder?
+    }
   end
 
   def update
@@ -98,13 +134,13 @@ class ScriptsController < ApplicationController
 
     script = Script.get_from_cache(params[:script_id])
 
-    render 'levels/instructions', locals: {stages: script.stages}
+    render 'levels/instructions', locals: {stages: script.lessons}
   end
 
   private
 
   def set_script_file
-    @script_file = ScriptDSL.serialize_stages(@script)
+    @script_file = ScriptDSL.serialize_lesson_groups(@script)
   end
 
   def rake
@@ -138,26 +174,33 @@ class ScriptsController < ApplicationController
     h = params.permit(
       :visible_to_teachers,
       :curriculum_umbrella,
+      :family_name,
+      :version_year,
       :project_sharing,
       :login_required,
-      :hideable_stages,
+      :hideable_stages, # TODO: remove once corresponding js change is deployed and no longer cached
+      :hideable_lessons,
       :curriculum_path,
       :professional_learning_course,
       :peer_reviews_to_complete,
       :wrapup_video,
       :student_detail_progress_view,
       :project_widget_visible,
-      :exclude_csf_column_in_legend,
-      :stage_extras_available,
+      :stage_extras_available, # TODO: remove once corresponding js change is deployed and no longer cached
+      :lesson_extras_available,
       :has_verified_resources,
       :has_lesson_plan,
+      :tts,
+      :is_stable,
       :script_announcements,
       :pilot_experiment,
+      :editor_experiment,
       resourceTypes: [],
       resourceLinks: [],
       project_widget_types: [],
       supported_locales: [],
     ).to_h
+    h[:lesson_extras_available] ||= h[:stage_extras_available] # TODO: remove once corresponding js change is deployed and no longer cached
     h[:peer_reviews_to_complete] = h[:peer_reviews_to_complete].to_i
     h[:hidden] = !h[:visible_to_teachers]
     h[:script_announcements] = JSON.parse(h[:script_announcements]) if h[:script_announcements]
